requirements/base.txt
# Core Django
Django==5.1.4  # Updated for security patches (CVE fixes)
djangorestframework==3.15.2
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Redis & Caching
redis==5.0.8
django-redis==5.4.0

# Celery
celery==5.4.0  # Stable LTS version (5.5.x requires testing)
django-celery-beat==2.7.0
flower==2.0.1

# API Documentation
drf-spectacular==0.27.2

# CORS
django-cors-headers==4.3.1

# Persian/Jalali Support
jdatetime==5.0.0

# SMS (Kavenegar)
kavenegar==1.1.2

# Utilities
python-decouple==3.8
requests==2.32.3  # Latest stable
Pillow==10.4.0

# Payment & Validation
pydantic==2.9.2
tenacity==9.0.0
python-dateutil==2.9.0
